backend/pyproject.toml
[project]
name = "spatial-jobs-index-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "geojson>=3.2.0",
    "python-dotenv>=1.0.1",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.0",
    "sqlalchemy[mypy]>=2.0.0",
    "psycopg2-binary>=2.9.10",
    "geoalchemy2>=0.15.0",
    "pyghtcast",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.13",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "python-lsp-server>=1.12.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "faker>=30.0.0",
    "factory-boy>=3.3.0",
]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
check_untyped_defs = false
disallow_untyped_defs = false
warn_return_any = false
warn_unused_configs = true
strict_optional = true

[tool.uv.sources]
pyghtcast = { git = "https://github.com/Dallas-College-LMIC/pyghtcast" }
